Remove internal endpoint for object counts in downstream clusters
